# packages/dashboard-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Along Dashboard & Knowledge Base API",
    "description": "Type-safe OpenAPI REST & SSE service for Along Protocol entities, dependency DAGs, and Knowledge Base search.",
    "version": "2.0.8"
  },
  "paths": {
...
    "/api/entities/issues": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "List all issues",
        "operationId": "listIssues",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by status (open, in-progress, blocked, done)",
              "title": "Status"
            },
            "description": "Filter by status (open, in-progress, blocked, done)"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by type (feat, bug, debt, task, docs)",
              "title": "Type"
            },
            "description": "Filter by type (feat, bug, debt, task, docs)"
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by priority (critical, high, medium, low)",
              "title": "Priority"
            },
            "description": "Filter by priority (critical, high, medium, low)"
          },
          {
            "name": "milestone",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by milestone slug",
              "title": "Milestone"
            },
            "description": "Filter by milestone slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/IssueSchema"
                  },
                  "title": "Response Listissues"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/entities/issues/{issue_id}": {
      "get": {
        "tags": [
          "Entities"
        ],
        "summary": "Get issue details",
        "operationId": "getIssue",
        "parameters": [
          {
            "name": "issue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Issue Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssueSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kb": {
      "get": {
        "tags": [
          "Knowledge Base"
        ],
        "summary": "List all Knowledge Base articles",
        "operationId": "listArticles",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category type",
              "title": "Type"
            },
            "description": "Filter by category type"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag",
              "title": "Tag"
            },
            "description": "Filter by tag"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/KBArticleSchema"
                  },
                  "title": "Response Listarticles"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kb/search": {
      "get": {
        "tags": [
          "Knowledge Base"
        ],
        "summary": "Search Knowledge Base and entities",
        "operationId": "searchKb",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Search query string",
              "default": "",
              "title": "Q"
            },
            "description": "Search query string"
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by tag",
              "title": "Tag"
            },
            "description": "Filter by tag"
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity type (kb, issue, decision, session)",
              "title": "Type"
            },
            "description": "Filter by entity type (kb, issue, decision, session)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/kb/{slug}": {
      "get": {
        "tags": [
          "Knowledge Base"
        ],
        "summary": "Get single Knowledge Base article",
        "operationId": "getArticle",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KBArticleSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IssueSchema": {
        "properties": {
          "protocol": {
            "type": "string",
            "title": "Protocol",
            "description": "Protocol identifier",
            "default": "along"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Unique kebab-case slug of the entity"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Canonical entity identifier, e.g. feat--login"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Human readable title"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body",
            "description": "Markdown body content"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path",
            "description": "Relative file path in repository"
          },
          "type": {
            "type": "string",
            "enum": [
              "feat",
              "bug",
              "debt",
              "task",
              "docs"
            ],
            "title": "Type",
            "description": "Issue type",
            "default": "feat"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "in-progress",
              "blocked",
              "done"
            ],
            "title": "Status",
            "description": "Current lifecycle status",
            "default": "open"
          },
          "priority": {
            "type": "string",
            "enum": [
              "critical",
              "high",
              "medium",
              "low"
            ],
            "title": "Priority",
            "description": "Priority level",
            "default": "medium"
          },
          "created": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created",
            "description": "Creation date YYYY-MM-DD"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated",
            "description": "Last update date YYYY-MM-DD"
          },
          "completed": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed",
            "description": "Completion date YYYY-MM-DD"
          },
          "agent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent",
            "description": "Agent or author that worked on the issue"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Array of topical tags"
          },
          "milestone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Milestone",
            "description": "Milestone slug this issue belongs to"
          },
          "blocked_by": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Blocked By",
            "description": "List of blocking entity keys"
          },
          "related": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related",
            "description": "List of related entity keys"
          },
          "parent": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent",
            "description": "Parent epic/entity key"
          }
        },
        "type": "object",
        "required": [
          "slug",
          "id"
        ],
        "title": "IssueSchema",
        "description": "Schema for .along/ISSUES/<type>--<slug>.md."
      },
      "KBArticleSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique article ID, e.g. kb--architecture or docs--setup"
          },
          "slug": {
            "type": "string",
            "title": "Slug",
            "description": "Article slug or filename"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "Article title"
          },
          "type": {
            "type": "string",
            "enum": [
              "topic",
              "architecture",
              "domain-model",
              "setup-workflow",
              "index",
              "doc"
            ],
            "title": "Type",
            "description": "Article topic category",
            "default": "topic"
          },
          "created": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created",
            "description": "Creation date YYYY-MM-DD"
          },
          "updated": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated",
            "description": "Last update date YYYY-MM-DD"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "List of topic tags"
          },
          "file_path": {
            "type": "string",
            "title": "File Path",
            "description": "Relative path in repository"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "description": "Markdown body",
            "default": ""
          },
          "headings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Headings",
            "description": "Extracted Markdown headings"
          },
          "outgoing_links": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Outgoing Links",
            "description": "Outgoing wiki or entity links"
          },
          "incoming_links": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Incoming Links",
            "description": "Incoming links referencing this article"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "title",
          "file_path"
        ],
        "title": "KBArticleSchema",
        "description": "Schema for Knowledge Base article in .along/KB/ or docs/."
      },
...
      "SearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResultItem"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "query",
          "total"
        ],
        "title": "SearchResponse"
      },
      "SearchResultItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "snippet": {
            "type": "string",
            "title": "Snippet"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "default": 1.0
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "type",
          "snippet",
          "file_path"
        ],
        "title": "SearchResultItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}